refactor(*): move helper config files to own directory
NODE-1093
