+++ lib/runner.js
@@ -4,11 +4,11 @@ var path = require('path'),
     f = require('util').format,
     fs = require('fs'),
     Metamocha = require('metamocha').Metamocha,
-    Configuration = require('./test_configuration'),
+    Configuration = require('./config/test_configuration'),
     ServerManager = require('mongodb-topology-manager').Server,
     ReplSetManager = require('mongodb-topology-manager').ReplSet,
     ShardingManager = require('mongodb-topology-manager').Sharded,
-    setupShardedCluster = require('./sharding_test_topology').setupShardedCluster,
+    setupShardedCluster = require('./config/sharding_topology_setup').setupShardedCluster,
     NodeVersionFilter = require('./filters/node_version_filter'),
     MongoDBVersionFilter = require('./filters/mongodb_version_filter'),
     MongoDBTopologyFilter = require('./filters/mongodb_topology_filter'),
